Display permitted URLs on popup list view
Passed in the entire context object since it looked like we're
going to need all of the data for the popup configs. If we want to
we can always encapsulate more later.
I added a quick draft of displaying data in a list. Still awaiting
a save button that works.